requirements.txt
PyYAML==6.0
wandb==0.13.1
multimethod==1.8
python-dotenv==0.20.0
plotly==5.11.0
pandas==1.4.2
scikit-learn==1.1.3
scipy==1.7.3
plotly==5.11.0
numpy==1.22.2
h5py==3.7.0
trimesh==3.16.4
Cython==0.29.32
cutil==3.0.2